Fix: 소수점 타입 계산 수정

+++ src/main/java/ch01/Statement.java
@@ -20,10 +20,10 @@ public class Statement {
             volumeCredits = volumeCreditsFor(perf);
 
             // 청구 내역을 출력한다.
-            result += "    " + playFor(perf).name() + ": " + format.format(amountFor(perf) / 100) + " (" + perf.audience() + "석)\n";
+            result += "    " + playFor(perf).name() + ": " + format.format((double) amountFor(perf) / 100) + " (" + perf.audience() + "석)\n";
             totalAmount += amountFor(perf);
         }
-        result += "총액: " + format.format(totalAmount/100) + "\n";
+        result += "총액: " + format.format((double) totalAmount / 100) + "\n";
         result += "적립 포인트: " + volumeCredits + "점\n";
         return result;
     }
